Made a very simple prototype.
I had a discussion with my tutor and came to the conclusion that we are overcomplicating our prototype. We need the minimal viable product before we start adding things like user interface and actual databases behind the scenes. I met with Jonathan and showed him a very simple prototype I made and he agreed with me. Here is a brief explanation of this prototype: The "database" is just a file of employee names. The program runs and asks for a login. If you successfully login, the only option you have is to print out all of the employees and then the program ends. This is all done in the command line. This leaves room for much more expansion in the future and gives us a good starting point.