Comment all of the code for the game.
This should help you understand the code faster. One thing to keep in mind as you work on this in the future. We want components to contain ONLY DATA and we want engine to contain ONLY FUNCTIONS. The components hold the data and the engines work on that data. If you are adding functions to a component or adding variables to an engine, you're probably doing something wrong. Always feel free to contact me for help.